backend/requirements.txt
# FastAPI and Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.3.0

# Database
sqlalchemy==2.0.36
alembic==1.13.2
psycopg2-binary==2.9.10
asyncpg==0.29.0

# Authentication
authlib==1.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Background Tasks
celery==5.4.0
redis==5.2.0

# HTTP Client
httpx==0.27.2
requests==2.32.3

# PDF Generation
weasyprint==61.2

# AWS (for S3 reports)
boto3==1.35.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.8.0
flake8==7.1.1
mypy==1.11.1



